Fix crash after going training -> css -> training
What was happening was that the entire program was synced with the game's
internal frame counter, and that counter only goes up outside of menus,
including the character select screen. This means that the logic to set
Action_Handler::is_save_flag to false whenever the game goes to CSS would
never work, because to the program, the game was never in the CSS.

But now that logic runs independently from the game's frame counter.